+++ morpheus/resource_mvm_instance.go
@@ -2,8 +2,10 @@ package morpheus
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"log"
+	"strconv"
 	"time"
 
 	"github.com/gomorpheus/morpheus-go-sdk"
@@ -12,13 +14,13 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func resourceVsphereInstance() *schema.Resource {
+func resourceMVMInstance() *schema.Resource {
 	return &schema.Resource{
 		Description:   "Provides a Morpheus instance resource.",
-		CreateContext: resourceVsphereInstanceCreate,
-		ReadContext:   resourceVsphereInstanceRead,
-		UpdateContext: resourceVsphereInstanceUpdate,
-		DeleteContext: resourceVsphereInstanceDelete,
+		CreateContext: resourceMVMInstanceCreate,
+		ReadContext:   resourceMVMInstanceRead,
+		UpdateContext: resourceMVMInstanceUpdate,
+		DeleteContext: resourceMVMInstanceDelete,
 		Timeouts: &schema.ResourceTimeout{
 			Create: schema.DefaultTimeout(45 * time.Minute),
 			Read:   schema.DefaultTimeout(5 * time.Minute),
@@ -33,6 +35,12 @@ func resourceVsphereInstance() *schema.Resource {
 			},
 			"name": {
 				Description: "The name of the instance",
+				Type:        schema.TypeString,
+				Optional:    true,
+				Computed:    true,
+			},
+			"display_name": {
+				Description: "The display name of the instance",
 				Type:        schema.TypeString,
 				Optional:    true,
 				Computed:    true,
@@ -55,13 +63,13 @@ func resourceVsphereInstance() *schema.Resource {
 				Required:    true,
 			},
 			"instance_type_id": {
-				Description: "The type of instance to provision",
+				Description: "The ID of the instance type to provision the instance from",
 				Type:        schema.TypeInt,
 				ForceNew:    true,
 				Required:    true,
 			},
 			"instance_layout_id": {
-				Description: "The layout to provision the instance from",
+				Description: "The ID of the instance layout to provision the instance from",
 				Type:        schema.TypeInt,
 				ForceNew:    true,
 				Required:    true,
@@ -69,12 +77,12 @@ func resourceVsphereInstance() *schema.Resource {
 			"plan_id": {
 				Description: "The service plan associated with the instance",
 				Type:        schema.TypeInt,
-				ForceNew:    true,
 				Required:    true,
 			},
-			"resource_pool_id": {
-				Description: "The ID of the resource pool to provision the instance to",
-				Type:        schema.TypeInt,
+			"resource_pool_name": {
+				Description: "The name of the resource pool (cluster) to provision the instance to",
+				Type:        schema.TypeString,
+				ForceNew:    true,
 				Optional:    true,
 				Computed:    true,
 			},
@@ -135,7 +143,14 @@ func resourceVsphereInstance() *schema.Resource {
 				Computed:    true,
 			},
 			"user_group_id": {
-				Description: "The id of the user group associated with the instance",
+				Description: "The ID of the user group associated with the instance",
+				Type:        schema.TypeInt,
+				ForceNew:    true,
+				Optional:    true,
+				Computed:    true,
+			},
+			"image_id": {
+				Description: "The ID of the image associated with the instance (Only neccessary when using the default MVM instance type that requires specifying a virtual image)",
 				Type:        schema.TypeInt,
 				ForceNew:    true,
 				Optional:    true,
@@ -143,6 +158,13 @@ func resourceVsphereInstance() *schema.Resource {
 			},
 			"asset_tag": {
 				Description: "The asset tag associated with the instance",
+				Type:        schema.TypeString,
+				ForceNew:    true,
+				Optional:    true,
+				Computed:    true,
+			},
+			"qemu_arguments": {
+				Description: "The qemu arguments to add to the instance",
 				Type:        schema.TypeString,
 				ForceNew:    true,
 				Optional:    true,
@@ -156,7 +178,14 @@ func resourceVsphereInstance() *schema.Resource {
 				Computed:    true,
 			},
 			"nested_virtualization": {
-				Description: "Whether to skip configuration of nested virtualization",
+				Description: "Whether to enable nested virtualization",
+				Type:        schema.TypeBool,
+				ForceNew:    true,
+				Optional:    true,
+				Computed:    true,
+			},
+			"attach_virtio_drivers": {
+				Description: "Whether to attach the virtio drivers to the instance",
 				Type:        schema.TypeBool,
 				ForceNew:    true,
 				Optional:    true,
@@ -191,12 +220,18 @@ func resourceVsphereInstance() *schema.Resource {
 					},
 				},
 			},
-			"volumes": {
+			"storage_volume": {
 				Description: "The instance volumes to create",
 				Type:        schema.TypeList,
 				Optional:    true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
+						"uuid": {
+							Description: "The storage volume uuid",
+							Type:        schema.TypeString,
+							Optional:    true,
+							Computed:    true,
+						},
 						"root": {
 							Description: "Whether the volume is the root volume of the instance",
 							Type:        schema.TypeBool,
@@ -213,14 +248,8 @@ func resourceVsphereInstance() *schema.Resource {
 							Optional:    true,
 							Computed:    true,
 						},
-						"size_id": {
-							Description: "The ID of an existing LV to assign to the instance",
-							Type:        schema.TypeInt,
-							Optional:    true,
-							Computed:    true,
-						},
 						"storage_type": {
-							Description: "The ID of the LV type",
+							Description: "The storage volume type ID",
 							Type:        schema.TypeInt,
 							Optional:    true,
 							Computed:    true,
@@ -234,7 +263,7 @@ func resourceVsphereInstance() *schema.Resource {
 					},
 				},
 			},
-			"interfaces": {
+			"network_interface": {
 				Description: "The instance network interfaces to create",
 				Type:        schema.TypeList,
 				Optional:    true,
@@ -246,20 +275,22 @@ func resourceVsphereInstance() *schema.Resource {
 							Optional:    true,
 							Computed:    true,
 						},
+						/* AWAITING PLATFORM SUPPORT
 						"network_group": {
 							Description: "Whether the network id provided is for a network group or not",
 							Type:        schema.TypeBool,
 							Optional:    true,
 							Computed:    true,
 						},
+						*/
 						"ip_address": {
-							Description: "",
+							Description: "The IP address to assign to the instance",
 							Type:        schema.TypeString,
 							Optional:    true,
 							Computed:    true,
 						},
 						"ip_mode": {
-							Description: "",
+							Description: "The IP address assignment mode",
 							Type:        schema.TypeString,
 							Optional:    true,
 							Computed:    true,
@@ -273,6 +304,12 @@ func resourceVsphereInstance() *schema.Resource {
 					},
 				},
 			},
+			"primary_ip_address": {
+				Description: "The instance primary IP address",
+				Type:        schema.TypeString,
+				Optional:    true,
+				Computed:    true,
+			},
 		},
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
@@ -280,7 +317,7 @@ func resourceVsphereInstance() *schema.Resource {
 	}
 }
 
-func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+func resourceMVMInstanceCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*morpheus.Client)
 
 	// Warning or errors can be collected in a slice type
@@ -289,6 +326,7 @@ func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, 
 	group := d.Get("group_id").(int)
 	cloud := d.Get("cloud_id").(int)
 	name := d.Get("name").(string)
+	displayName := d.Get("display_name").(string)
 
 	// Service Plan
 	planResp, err := client.GetPlan(int64(d.Get("plan_id").(int)), &morpheus.Request{})
@@ -318,13 +356,26 @@ func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, 
 	config := make(map[string]interface{})
 
 	// Resource Pool
-	resourcePoolResp, err := client.GetResourcePool(int64(cloud), int64(d.Get("resource_pool_id").(int)), &morpheus.Request{})
+	resourcePoolResp, err := client.Execute(&morpheus.Request{
+		Method:      "GET",
+		Path:        fmt.Sprintf("/api/options/zonePools?layoutId=%d", d.Get("instance_layout_id").(int)),
+		QueryParams: map[string]string{},
+	})
 	if err != nil {
 		diag.FromErr(err)
 	}
-	resourcePoolResult := resourcePoolResp.Result.(*morpheus.GetResourcePoolResult)
-	resourcePool := resourcePoolResult.ResourcePool
-	config["resourcePoolId"] = resourcePool.ID
+
+	var itemResponsePayload ResourcePoolOptions
+	json.Unmarshal(resourcePoolResp.Body, &itemResponsePayload)
+	var resourcePoolId int
+	for _, v := range itemResponsePayload.Data {
+		if v.ProviderType == "mvm" && v.Name == d.Get("resource_pool_name").(string) {
+			resourcePoolId = v.Id
+		}
+	}
+
+	config["resourcePoolId"] = resourcePoolId
+	config["poolProviderType"] = "mvm"
 
 	// Custom Options
 	if d.Get("custom_options") != nil {
@@ -344,6 +395,11 @@ func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, 
 		config["smbiosAssetTag"] = d.Get("asset_tag").(string)
 	}
 
+	// Image ID
+	if d.Get("image_id") != nil {
+		config["imageId"] = d.Get("image_id").(int)
+	}
+
 	// Skip Agent Install
 	config["noAgent"] = d.Get("skip_agent_install").(bool)
 
@@ -351,8 +407,9 @@ func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, 
 	config["nestedVirtualization"] = d.Get("nested_virtualization").(bool)
 
 	instancePayload := map[string]interface{}{
-		"name": name,
-		"type": instanceTypeCode,
+		"name":        name,
+		"displayName": displayName,
+		"type":        instanceTypeCode,
 		"site": map[string]interface{}{
 			"id": group,
 		},
@@ -433,13 +490,13 @@ func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, 
 	}
 
 	// Network Interfaces
-	if d.Get("interfaces") != nil {
-		payload["networkInterfaces"] = parseNetworkInterfaces(d.Get("interfaces").([]interface{}))
+	if d.Get("network_interface") != nil {
+		payload["networkInterfaces"] = generateMVMInstanceCreatePayload(d.Get("network_interface").([]interface{}))
 	}
 
 	// Volumes
-	if d.Get("volumes") != nil {
-		payload["volumes"] = parseStorageVolumes(d.Get("volumes").([]interface{}))
+	if d.Get("storage_volume") != nil {
+		payload["volumes"] = parseStorageVolumes(d.Get("storage_volume").([]interface{}))
 	}
 
 	req := &morpheus.Request{Body: payload}
@@ -451,6 +508,7 @@ func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, 
 	log.Printf("API RESPONSE: %s", resp)
 	result := resp.Result.(*morpheus.CreateInstanceResult)
 	instance := result.Instance
+	instanceStatus := "provisioning"
 
 	stateConf := &resource.StateChangeConf{
 		Pending: []string{"provisioning", "starting", "stopping", "pending"},
@@ -462,12 +520,13 @@ func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, 
 			}
 			result := instanceDetails.Result.(*morpheus.GetInstanceResult)
 			instance := result.Instance
+			instanceStatus = instance.Status
 			return result, instance.Status, nil
 		},
 		Timeout:      3 * time.Hour,
 		MinTimeout:   1 * time.Minute,
-		Delay:        3 * time.Minute,
-		PollInterval: 1 * time.Minute,
+		Delay:        2 * time.Minute,
+		PollInterval: 30 * time.Second,
 	}
 
 	// Wait, catching any errors
@@ -478,11 +537,17 @@ func resourceVsphereInstanceCreate(ctx context.Context, d *schema.ResourceData, 
 
 	// Successfully created resource, now set id
 	d.SetId(int64ToString(instance.ID))
-	resourceVsphereInstanceRead(ctx, d, meta)
+	resourceMVMInstanceRead(ctx, d, meta)
+
+	// Fail the instance deployment if the
+	// instance status is in a failed state
+	if instanceStatus == "failed" {
+		return diag.Errorf("error creating instance: failed to create server")
+	}
 	return diags
 }
 
-func resourceVsphereInstanceRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+func resourceMVMInstanceRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*morpheus.Client)
 
 	// Warning or errors can be collected in a slice type
@@ -529,10 +594,21 @@ func resourceVsphereInstanceRead(ctx context.Context, d *schema.ResourceData, me
 	d.Set("instance_type_id", instance.InstanceType.ID)
 	d.Set("instance_layout_id", instance.Layout.ID)
 	d.Set("plan_id", instance.Plan.ID)
-	d.Set("resource_pool_id", instance.Config["resourcePoolId"])
 	d.Set("environment", instance.Environment)
 	d.Set("labels", instance.Labels)
-	d.Set("evar", instance.EnvironmentVariables)
+
+	var evars []map[string]interface{}
+	for i := 0; i < len(instance.EnvironmentVariables); i++ {
+		evar := instance.EnvironmentVariables[i]
+		row := make(map[string]interface{})
+		row["name"] = evar.Name
+		row["value"] = fmt.Sprintf("%v", evar.Value)
+		row["export"] = evar.Export
+		row["masked"] = evar.Masked
+		evars = append(evars, row)
+	}
+	d.Set("evar", evars)
+
 	// Tags
 	tags := make(map[string]interface{})
 	if instance.Tags != nil {
@@ -546,6 +622,7 @@ func resourceVsphereInstanceRead(ctx context.Context, d *schema.ResourceData, me
 		}
 	}
 	d.Set("tags", tags)
+
 	if instance.Config["userGroup"] != nil {
 		userGroup := instance.Config["userGroup"].(map[string]interface{})
 		d.Set("user_group_id", userGroup["id"])
@@ -560,14 +637,76 @@ func resourceVsphereInstanceRead(ctx context.Context, d *schema.ResourceData, me
 	}
 	d.Set("custom_options", instance.Config["customOptions"])
 	d.Set("domain_id", instance.NetworkDomain.Id)
+	d.Set("primary_ip_address", instance.ConnectionInfo[0].Ip)
+
+	var volumes []map[string]interface{}
+	// iterate over the array of svcports
+	for i := 0; i < len(instance.Volumes); i++ {
+		row := make(map[string]interface{})
+		volume := instance.Volumes[i]
+		row["uuid"] = volume.Uuid
+		row["root"] = volume.RootVolume.(bool)
+		row["name"] = volume.Name
+		row["size"] = volume.Size.(float64)
+		row["storage_type"] = volume.StorageType.(float64)
+		n, err := strconv.Atoi(volume.DatastoreId)
+		if err != nil {
+			log.Printf("API FAILURE: %s - %s", resp, err)
+			return diag.FromErr(err)
+		}
+		row["datastore_id"] = n
+		volumes = append(volumes, row)
+	}
+	d.Set("storage_volume", volumes)
+
+	var networkInterfaces []map[string]interface{}
+	// iterate over the array of svcports
+	for i := 0; i < len(instance.Interfaces); i++ {
+		row := make(map[string]interface{})
+		networkInterface := instance.Interfaces[i]
+		row["network_id"] = int(networkInterface.Network.ID)
+		row["ip_address"] = networkInterface.IpAddress
+		row["ip_mode"] = networkInterface.IpMode
+		row["network_interface_type_id"] = networkInterface.NetworkInterfaceTypeId
+		networkInterfaces = append(networkInterfaces, row)
+	}
+	d.Set("network_interface", networkInterfaces)
+
 	return diags
 }
 
-func resourceVsphereInstanceUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+func resourceMVMInstanceUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*morpheus.Client)
 	id := d.Id()
-	name := d.Get("name").(string)
-	description := d.Get("description").(string)
+
+	instanceGetResp, err := client.GetInstance(toInt64(id), &morpheus.Request{})
+	result := instanceGetResp.Result.(*morpheus.GetInstanceResult)
+	fetchInstance := result.Instance
+
+	var instance *morpheus.Instance
+
+	instancePayload := make(map[string]interface{})
+	baseInstanceConfigUpdate := false
+
+	if d.HasChange("name") {
+		instancePayload["name"] = d.Get("name").(string)
+		baseInstanceConfigUpdate = true
+	}
+
+	if d.HasChange("description") {
+		instancePayload["description"] = d.Get("description").(string)
+		baseInstanceConfigUpdate = true
+	}
+
+	if d.HasChange("display_name") {
+		instancePayload["displayName"] = d.Get("display_name").(string)
+		baseInstanceConfigUpdate = true
+	}
+
+	if d.HasChange("environment") {
+		instancePayload["instanceContext"] = d.Get("environment").(string)
+		baseInstanceConfigUpdate = true
+	}
 
 	// Tags
 	var tags []map[string]interface{}
@@ -579,6 +718,8 @@ func resourceVsphereInstanceUpdate(ctx context.Context, d *schema.ResourceData, 
 			tag["value"] = value.(string)
 			tags = append(tags, tag)
 		}
+		baseInstanceConfigUpdate = true
+		instancePayload["tags"] = tags
 	}
 	config := make(map[string]interface{})
 
@@ -590,34 +731,92 @@ func resourceVsphereInstanceUpdate(ctx context.Context, d *schema.ResourceData, 
 			customOptions[key] = value.(string)
 		}
 		config["customOptions"] = customOptions
+		baseInstanceConfigUpdate = true
 	}
 
-	instancePayload := map[string]interface{}{
-		"name":            name,
-		"description":     description,
-		"labels":          d.Get("labels"),
-		"tags":            tags,
-		"instanceContext": d.Get("environment"),
-		"config":          config,
+	if d.HasChanges("labels") {
+		instancePayload["labels"] = d.Get("labels")
+		baseInstanceConfigUpdate = true
 	}
-	payload := map[string]interface{}{
-		"instance": instancePayload,
+
+	if baseInstanceConfigUpdate {
+		payload := map[string]interface{}{
+			"instance": instancePayload,
+		}
+		req := &morpheus.Request{Body: payload}
+		resp, err := client.UpdateInstance(toInt64(id), req)
+		if err != nil {
+			log.Printf("API FAILURE: %s - %s", resp, err)
+			return diag.FromErr(err)
+		}
+		log.Printf("API RESPONSE: %s", resp)
+		result := resp.Result.(*morpheus.UpdateInstanceResult)
+		instance = result.Instance
 	}
-	req := &morpheus.Request{Body: payload}
-	resp, err := client.UpdateInstance(toInt64(id), req)
+
+	// Resize Instance Configuration
+	var resizeChanges bool = false
+
+	resizeInstancePayload := make(map[string]interface{})
+
+	if d.HasChange("plan_id") {
+		resizeChanges = true
+		resizeInstancePayload["instance"] = map[string]interface{}{
+			"plan": map[string]interface{}{
+				"id": d.Get("plan_id"),
+			},
+		}
+	}
+
+	if d.HasChange("storage_volume") || d.HasChange("network_interface") {
+		resizeChanges = true
+		resizeInstancePayload["networkInterfaces"] = parseMVMNetworkInterfaces(d.Get("network_interface").([]interface{}), fetchInstance.Interfaces)
+		resizeInstancePayload["volumes"] = parseMVMStorageVolumes(d.Get("storage_volume").([]interface{}), fetchInstance.Volumes)
+		resizeInstancePayload["deleteOriginalVolumes"] = true
+	}
+
+	// Check if plan, storage, or nics have changed
+	if resizeChanges {
+		resizeReq := &morpheus.Request{Body: resizeInstancePayload}
+		resizeResp, err := client.ResizeInstance(toInt64(id), resizeReq)
+		if err != nil {
+			log.Printf("API FAILURE: %s - %s", resizeResp, err)
+			return diag.FromErr(err)
+		}
+		result := resizeResp.Result.(*morpheus.UpdateInstanceResult)
+		instance = result.Instance
+	}
+
+	stateConf := &resource.StateChangeConf{
+		Pending: []string{"resizing", "pending"},
+		Target:  []string{"running"},
+		Refresh: func() (interface{}, string, error) {
+			instanceDetails, err := client.GetInstance(toInt64(id), &morpheus.Request{})
+			if err != nil {
+				return "", "", err
+			}
+			result := instanceDetails.Result.(*morpheus.GetInstanceResult)
+			instance := result.Instance
+			return result, instance.Status, nil
+		},
+		Timeout:      30 * time.Minute,
+		MinTimeout:   1 * time.Minute,
+		Delay:        1 * time.Minute,
+		PollInterval: 30 * time.Second,
+	}
+
+	// Wait, catching any errors
+	_, err = stateConf.WaitForStateContext(ctx)
 	if err != nil {
-		log.Printf("API FAILURE: %s - %s", resp, err)
-		return diag.FromErr(err)
+		return diag.Errorf("error updating instance: %s", err)
 	}
-	log.Printf("API RESPONSE: %s", resp)
-	result := resp.Result.(*morpheus.UpdateInstanceResult)
-	instance := result.Instance
+
 	// Successfully updated resource, now set id
 	d.SetId(int64ToString(instance.ID))
-	return resourceVsphereInstanceRead(ctx, d, meta)
+	return resourceMVMInstanceRead(ctx, d, meta)
 }
 
-func resourceVsphereInstanceDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+func resourceMVMInstanceDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*morpheus.Client)
 
 	// Warning or errors can be collected in a slice type
@@ -641,24 +840,60 @@ func resourceVsphereInstanceDelete(ctx context.Context, d *schema.ResourceData, 
 		}
 	}
 	log.Printf("API RESPONSE: %s", resp)
+
+	stateConf := &resource.StateChangeConf{
+		Pending: []string{"removing", "pendingRemoval", "stopping", "pending"},
+		Target:  []string{"removed"},
+		Refresh: func() (interface{}, string, error) {
+			instanceDetails, err := client.GetInstance(toInt64(id), &morpheus.Request{})
+			if instanceDetails.StatusCode == 404 {
+				return "", "removed", nil
+			}
+			if err != nil {
+				return "", "", err
+			}
+			result := instanceDetails.Result.(*morpheus.GetInstanceResult)
+			instance := result.Instance
+			return result, instance.Status, nil
+		},
+		Timeout:      30 * time.Minute,
+		MinTimeout:   1 * time.Minute,
+		Delay:        1 * time.Minute,
+		PollInterval: 30 * time.Second,
+	}
+
+	// Wait, catching any errors
+	_, err = stateConf.WaitForStateContext(ctx)
+	if err != nil {
+		return diag.Errorf("error deleting instance: %s", err)
+	}
+
 	d.SetId("")
 	return diags
 }
 
-func parseNetworkInterfaces(interfaces []interface{}) []map[string]interface{} {
+type ResourcePoolOptions struct {
+	Success bool `json:"success"`
+	Data    []struct {
+		Id           int    `json:"id"`
+		Name         string `json:"name"`
+		IsGroup      bool   `json:"isGroup"`
+		Group        string `json:"group"`
+		IsDefault    bool   `json:"isDefault"`
+		Type         string `json:"type"`
+		ProviderType string `json:"providerType"`
+		Value        string `json:"value"`
+	} `json:"data"`
+}
+
+func generateMVMInstanceCreatePayload(interfaces []interface{}) []map[string]interface{} {
 	var networkInterfaces []map[string]interface{}
 	for i := 0; i < len(interfaces); i++ {
 		row := make(map[string]interface{})
 		item := (interfaces)[i].(map[string]interface{})
 		if item["network_id"] != nil {
-			if item["network_group"].(bool) {
-				row["network"] = map[string]interface{}{
-					"id": fmt.Sprintf("networkGroup-%d", item["network_id"].(int)),
-				}
-			} else {
-				row["network"] = map[string]interface{}{
-					"id": fmt.Sprintf("network-%d", item["network_id"].(int)),
-				}
+			row["network"] = map[string]interface{}{
+				"id": fmt.Sprintf("network-%d", item["network_id"].(int)),
 			}
 		}
 		if item["ip_address"] != nil {
@@ -675,14 +910,63 @@ func parseNetworkInterfaces(interfaces []interface{}) []map[string]interface{} {
 	return networkInterfaces
 }
 
-func parseStorageVolumes(volumes []interface{}) []map[string]interface{} {
+func parseMVMNetworkInterfaces(interfaces []interface{}, existingInterfaces []morpheus.NetworkInterface) []map[string]interface{} {
+	var networkInterfaces []map[string]interface{}
+	//var existingNetworkInterfaceIDs []string
+
+	for i := 0; i < len(interfaces); i++ {
+		row := make(map[string]interface{})
+		item := (interfaces)[i].(map[string]interface{})
+
+		for _, nic := range existingInterfaces {
+			if nic.ID == item["id"] {
+				row["id"] = nic.ID
+			}
+		}
+		if item["network_id"] != nil {
+			row["network"] = map[string]interface{}{
+				"id": fmt.Sprintf("network-%d", item["network_id"].(int)),
+			}
+		}
+		if item["ip_address"] != nil {
+			row["ipAddress"] = item["ip_address"] //.(string)
+		}
+		if item["ip_mode"] != nil {
+			row["ipMode"] = item["ip_mode"] // .(string)
+		}
+		if item["network_interface_type_id"] != nil {
+			row["networkInterfaceTypeId"] = item["network_interface_type_id"] //.(int)
+		}
+		networkInterfaces = append(networkInterfaces, row)
+	}
+	return networkInterfaces
+}
+
+func parseMVMStorageVolumes(volumes []interface{}, existingVolumes []morpheus.StorageVolume) []map[string]interface{} {
 	var storageVolumes []map[string]interface{}
+	var existingVolumeUUIDs []string
+
+	// Iterate through the existing storage volumes and fetch the UUID to identify
+	// which defined volumes already exist on the instance
+	for _, vol := range existingVolumes {
+		existingVolumeUUIDs = append(existingVolumeUUIDs, vol.Uuid)
+	}
+
 	for i := 0; i < len(volumes); i++ {
 		row := make(map[string]interface{})
 		item := (volumes)[i].(map[string]interface{})
-		if item["id"] != nil {
-			row["id"] = item["id"]
+
+		for _, vol := range existingVolumes {
+			if vol.Uuid == item["uuid"] {
+				row["id"] = vol.ID
+			}
 		}
+
+		// If the defined disk does not already existing
+		if !containsString(existingVolumeUUIDs, item["uuid"].(string)) {
+			row["id"] = -1
+		}
+
 		if item["root"] != nil {
 			row["rootVolume"] = item["root"]
 		}
@@ -691,9 +975,6 @@ func parseStorageVolumes(volumes []interface{}) []map[string]interface{} {
 		}
 		if item["size"] != nil {
 			row["size"] = item["size"] // .(int)
-		}
-		if item["size_id"] != nil {
-			row["sizeId"] = item["size_id"] // .(int)
 		}
 		if item["storage_type"] != nil {
 			row["storageType"] = item["storage_type"] // .(int)